Close co-occurrences of two keywords in a Gallicagram copus

Retrieves the proportion of close co-occurrences of two keywords in one of the corpora (historical press, Gallica books, Le Monde newspaper) by year, month or day.

gallicagram_cooccur(
  keyword_1,
  keyword_2,
  corpus = "lemonde",
  from = 1945,
  to = 2022,
  resolution = "monthly"
)

Arguments

keyword_1

A character string. One of the two keywords to search.

keyword_2

A character string. The other keyword to search.

corpus

A character string. The corpus to search. Takes the following values: "press" for historical press, "books" for Gallica books, "lemonde" for Le Monde newspaper articles.

from

An integer. Starting year.

to

An integer. End year.

resolution

A character string. For press and lemonde can be either "yearly" or "monthly". For books can only be "yearly".

Value

A tibble. With the keyword, the number of occurrences (n_occur) or co-occurrences (n_cooccur), the total number of ngrams over the period of a given observation (n_grams or n_ngrams), the proportion of occurrences or co-occurrences of the keyword(s) over the period of a given observation (prop_occcur or prop_coocccur), the date at the beginning of the period of a given observation (date), the corpus, the resolution, the year and potentially the month and day of the observation.

Details

Close co-occurrences correspond to the number of 3-grams (4-grams in the Le Monde corpus) that contain the two keywords.

This function corresponds to the Contain route of the API.

Examples

gallicagram_cooccur("président", "mauvais")
#> # A tibble: 936 × 10
#>    date       keywo…¹ keywo…² n_coo…³ n_ngr…⁴ prop_…⁵  year month corpus resol…⁶
#>    <date>     <chr>   <chr>     <dbl>   <int>   <dbl> <int> <int> <chr>  <chr>  
#>  1 1945-01-01 présid… mauvais       0  166925       0  1945     1 lemon… monthly
#>  2 1945-02-01 présid… mauvais       0  162907       0  1945     2 lemon… monthly
#>  3 1945-03-01 présid… mauvais       0  201083       0  1945     3 lemon… monthly
#>  4 1945-04-01 présid… mauvais       0  189895       0  1945     4 lemon… monthly
#>  5 1945-05-01 présid… mauvais       0  208666       0  1945     5 lemon… monthly
#>  6 1945-06-01 présid… mauvais       0  220081       0  1945     6 lemon… monthly
#>  7 1945-07-01 présid… mauvais       0  245302       0  1945     7 lemon… monthly
#>  8 1945-08-01 présid… mauvais       0  239788       0  1945     8 lemon… monthly
#>  9 1945-09-01 présid… mauvais       0  246492       0  1945     9 lemon… monthly
#> 10 1945-10-01 présid… mauvais       0  254720       0  1945    10 lemon… monthly
#> # … with 926 more rows, and abbreviated variable names ¹​keyword_1, ²​keyword_2,
#> #   ³​n_cooccur, ⁴​n_ngrams, ⁵​prop_cooccur, ⁶​resolution