Internal function to prepare the parameters to use them in the API call.

Checks potential errors, translates the parameters, set parameters for the API call.

prepare_param(keyword, corpus, from, to, resolution, n_of = "grams")

Arguments

keyword

A character string. Keyword to search. The largest number of words the keyword can contain can be found in the resolution column of the list_corpora dataset.

corpus

A character string. The corpus to search. The list of available corpora can be found in the list_corpora dataset.

from

An integer or "earliest". The starting year. If set to "earliest", use the earliest date at which the data is reliable for this corpus, as described in list_corpora.

to

An integer or "latest". The end year. If set to "latest", use the latest date at which the data is reliable for this corpus, as described in list_corpora.

resolution

A character string. Can only be "daily", "monthly" or "yearly". The finest available resolution for the corpus selected can be found in the resolution column of the list_corpora dataset.

n_of

A character string. The type of object to compute number of occurrences for. If set to "grams", will compute the number of "grams" that correspond to the keyword for the given period. If set to "article" (only available for lemonde), will compute the number of articles that contain the keyword for the given period.

Value

A list. Contains cleaned function parameters (`keyword`, `corpus`, `from`, `to` and `resolution`).