Internal function to tidy the output of the API call.

Turns the output into a tibble, renames the variables and add the input parameters to the output tibble.

tidy_gallicagram(data, corpus, resolution)

Arguments

data

A data frame produced by the Gallicagram API

corpus

A character string. The corpus to search. Takes the following values: "press" for historical press, "books" for Gallica books, "lemonde" for Le Monde newspaper articles.

resolution

A character string. For lemonde can be either "yearly", "monthly" or "daily". For press can be either "yearly" or "monthly". For books can only be "yearly".

Value

A tibble. With the keyword, the number of occurrences (n_occur) or co-occurrences (n_cooccur), the total number of ngrams over the period of a given observation (n_grams or n_ngrams), the proportion of occurrences or co-occurrences of the keyword(s) over the period of a given observation (prop_occcur or prop_coocccur), the date at the beginning of the period of a given observation (date), the source, the resolution, the year and potentially the month and day of the observation.