Occurrences of a keyword in a Gallicagram copus

Retrieves the proportion of occurrences of a keyword in one of the corpora (historical press, Gallica books, Le Monde newspaper) by year, month or day.

gallicagram(
  keyword,
  corpus = "lemonde",
  from = 1945,
  to = 2022,
  resolution = "monthly"
)

Arguments

keyword

A character string. Keyword to search. Can be up to a 3-gram in the "books" and "press" corpora and a 4-gram in the "lemonde" corpus.

corpus

A character string. The corpus to search. Takes the following values: "press" for historical press, "books" for Gallica books, "lemonde" for Le Monde newspaper articles.

from

An integer. Starting year.

to

An integer. End year.

resolution

A character string. For lemonde can be either "yearly", "monthly" or "daily". For press can be either "yearly" or "monthly". For books can only be "yearly".

Value

A tibble. With the keyword, the number of occurrences (n_occur) or co-occurrences (n_cooccur), the total number of ngrams over the period of a given observation (n_grams or n_ngrams), the proportion of occurrences or co-occurrences of the keyword(s) over the period of a given observation (prop_occcur or prop_coocccur), the date at the beginning of the period of a given observation (date), the corpus, the resolution, the year and potentially the month and day of the observation.

Details

This function corresponds to the Query route of the API.

Examples

gallicagram("président")
#> # A tibble: 936 × 9
#>    date       keyword   n_occur n_grams prop_occur  year month corpus  resolut…¹
#>    <date>     <chr>       <int>   <int>      <dbl> <int> <int> <chr>   <chr>    
#>  1 1945-01-01 président     306  262131   0.00117   1945     1 lemonde monthly  
#>  2 1945-02-01 président     248  256110   0.000968  1945     2 lemonde monthly  
#>  3 1945-03-01 président     327  313806   0.00104   1945     3 lemonde monthly  
#>  4 1945-04-01 président     565  299972   0.00188   1945     4 lemonde monthly  
#>  5 1945-05-01 président     406  327917   0.00124   1945     5 lemonde monthly  
#>  6 1945-06-01 président     462  341077   0.00135   1945     6 lemonde monthly  
#>  7 1945-07-01 président     633  387425   0.00163   1945     7 lemonde monthly  
#>  8 1945-08-01 président     740  376034   0.00197   1945     8 lemonde monthly  
#>  9 1945-09-01 président     401  382928   0.00105   1945     9 lemonde monthly  
#> 10 1945-10-01 président     583  396383   0.00147   1945    10 lemonde monthly  
#> # … with 926 more rows, and abbreviated variable name ¹​resolution