Word the most often associated with a n-gram in a Gallicagram corpus

Returns the word the most frequently following a given ngram over the period. It can also include the words that precede by setting after = FALSE.

gallicagram_associated(
  keyword,
  corpus = "lemonde",
  from = 1945,
  to = 2022,
  nb_associated_words = 20,
  after = TRUE,
  stopwords = NULL
)

Arguments

keyword

A character string. Keyword to search associations for.

corpus

A character string. The corpus to search. Takes the following values: "press" for historical press, "books" for Gallica books, "lemonde" for Le Monde newspaper articles.

from

A number. Starting year.

to

A number. End year.

nb_associated_words

An integer. The number of most frequently associated words to return.

after

A boolean. Whether to consider only words following the keyword and not those preceding. Set to TRUE by default.

stopwords

A character vector of stopwords to remove. Can for instance be lsa::stopwords_fr. Removed after withdrawing apostrophes and letters preceding them. If NULL does not remove any stopwords.

Value

A tibble. With the most frequent ngrams containing the keyword mentioned (ngram), the associated_word, and the number of occurrences over the period (nb_occur). It also returns the input parameters keyword, corpus, from and to.

Details

This function corresponds to the Joker route of the API, accessed through the 'Joker' function on the Gallicagram app and is analogous to the 'Joker' function on Ngram Viewer.

For instance "camarade" is often followed by "staline" or "khrouchtchev" in Le Monde. The function returns the most frequent ngrams of the form "camarade *". Setting after = FALSE also includes the most frequent ngrams of the form "* camarade".

The keyword can be a 2-gram in the "books" and "press" corpora and a 3-gram in the "lemonde" corpus. Searching the "press" corpus can require a long running time.

Examples

gallicagram_associated("camarade")
#> # A tibble: 20 × 7
#>    nb_occur ngram                 associated_word keyword  corpus   from    to
#>       <int> <chr>                 <chr>           <chr>    <chr>   <dbl> <dbl>
#>  1     2622 camarade de           de              camarade lemonde  1945  2022
#>  2      282 camarade qui          qui             camarade lemonde  1945  2022
#>  3      226 camarade staline      staline         camarade lemonde  1945  2022
#>  4      226 camarade et           et              camarade lemonde  1945  2022
#>  5      211 camarade khrouchtchev khrouchtchev    camarade lemonde  1945  2022
#>  6      181 camarade du           du              camarade lemonde  1945  2022
#>  7      115 camarade des          des             camarade lemonde  1945  2022
#>  8      110 camarade à            à               camarade lemonde  1945  2022
#>  9      104 camarade mao          mao             camarade lemonde  1945  2022
#> 10       99 camarade d'école      d'école         camarade lemonde  1945  2022
#> 11       89 camarade a            a               camarade lemonde  1945  2022
#> 12       80 camarade en           en              camarade lemonde  1945  2022
#> 13       72 camarade d'écurie     d'écurie        camarade lemonde  1945  2022
#> 14       68 camarade président    président       camarade lemonde  1945  2022
#> 15       68 camarade kim          kim             camarade lemonde  1945  2022
#> 16       66 camarade dans         dans            camarade lemonde  1945  2022
#> 17       62 camarade d'études     d'études        camarade lemonde  1945  2022
#> 18       59 camarade d'enfance    d'enfance       camarade lemonde  1945  2022
#> 19       57 camarade tito         tito            camarade lemonde  1945  2022
#> 20       55 camarade jean         jean            camarade lemonde  1945  2022