Words the most often associated with a keyword in a Gallicagram corpus

Returns the word the most frequently at a given number of words (distance) from the keyword.

gallicagram_associated(
  keyword,
  corpus = "lemonde",
  from = 1945,
  to = 2022,
  n_results = 20,
  distance = "max",
  stopwords = rallicagram::stopwords_gallica[1:500]
)

Arguments

keyword

A character string. Keyword to search. Can be up to a 3-gram in the "books" and "press" corpora and a 4-gram in the "lemonde" corpus.

corpus

A character string. The corpus to search. Takes the following values: "press" for historical press, "books" for Gallica books, "lemonde" for Le Monde newspaper articles.

from

An integer. Starting year.

to

An integer. End year.

n_results

An integer. The number of most frequently associated words to return. n_results can also be set to "all" to return all the available results.

distance

An integer or "max". The word distance to which look for words associated with the keyword. If equal to "max", will be 4 for the Le Monde corpus and 3 otherwise.

stopwords

A character vector of stopwords to remove. The default is the vector of the 500 most frequent words in the Gallica books dataset. We can change this number by passing stopwords_gallicca[1:300] (for instance, for the 300 most frequent) to the stopwords argument. Can also be lsa::stopwords_fr If NULL does not remove any stopwords.

Value

A tibble. Containing the words the most frequently associated with the keyword mentioned (associated_word), and the number of occurrences over the period (n_occur). It also returns the input parameters keyword, corpus, from and to.

Details

This functions calls the 'associated' route of the API.

Note that the API route does not allow to search for associated words after a punctuation mark. For instance, in "son camarade, le chasseur, a", the function will not count "chasseur" as a word associated with "camarade". The ngram "son camarade le chasseur" is not in the database. Thus, there might be less associated words at a longer distance as it increases the probability of ngrams to be excluded from the database.

Apostrophes and letters preceding them are withdrawn from the dataset.

Searching the "press" corpus can require a long running time.

Examples

gallicagram_associated("camarade")
#> # A tibble: 20 × 7
#>    associated_word n_occur keyword  corpus   from    to distance
#>    <chr>             <int> <chr>    <chr>   <dbl> <dbl> <chr>   
#>  1 ancien              480 camarade lemonde  1945  2022 max     
#>  2 classe              318 camarade lemonde  1945  2022 max     
#>  3 promotion           183 camarade lemonde  1945  2022 max     
#>  4 vieux               124 camarade lemonde  1945  2022 max     
#>  5 khrouchtchev        106 camarade lemonde  1945  2022 max     
#>  6 mao                  97 camarade lemonde  1945  2022 max     
#>  7 staline              96 camarade lemonde  1945  2022 max     
#>  8 qil                  85 camarade lemonde  1945  2022 max     
#>  9 parti                80 camarade lemonde  1945  2022 max     
#> 10 ami                  78 camarade lemonde  1945  2022 max     
#> 11 école                77 camarade lemonde  1945  2022 max     
#> 12 combat               72 camarade lemonde  1945  2022 max     
#> 13 toung                72 camarade lemonde  1945  2022 max     
#> 14 kim                  71 camarade lemonde  1945  2022 max     
#> 15 lycée                70 camarade lemonde  1945  2022 max     
#> 16 compagnie            67 camarade lemonde  1945  2022 max     
#> 17 club                 61 camarade lemonde  1945  2022 max     
#> 18 qun                  59 camarade lemonde  1945  2022 max     
#> 19 ancienne             58 camarade lemonde  1945  2022 max     
#> 20 ex                   51 camarade lemonde  1945  2022 max