Close co-occurrences of two keywords in a Gallicagram copus

Retrieves the proportion of close co-occurrences of two keywords in one of the corpora by year, month or day.

gallicagram_cooccur(
  keyword_1,
  keyword_2,
  corpus = "lemonde",
  from = 1945,
  to = 2022,
  resolution = "monthly",
  count_period = TRUE,
  n_of = "grams"
)

Arguments

keyword_1

A character string. One of the two keywords to search.

keyword_2

A character string. The other keyword to search.

corpus

A character string. The corpus to search. The list of available corpora can be found in the list_corpora dataset.

from

An integer. Starting year.

to

An integer. End year. The finest available resolution for the corpus selected can be found in the resolution column of the list_corpora dataset.

resolution

A character string. For press and lemonde can be either "yearly" or "monthly". For books can only be "yearly".

count_period

If FALSE, counts the co-occurrences of each phrase containing both keywords. If TRUE, returns the number of times both keywords co-occur in each resolution period.

Value

A tibble. With the keyword, the number of occurrences (n_occur) or co-occurrences (n_cooccur), the total number of ngrams over the period of a given observation (n_grams or n_ngrams), the proportion of occurrences or co-occurrences of the keyword(s) over the period of a given observation (prop_occcur or prop_coocccur), the date at the beginning of the period of a given observation (date), the corpus, the resolution, the year and potentially the month and day of the observation.

Details

Close co-occurrences correspond to the number of 3-grams (4-grams in the Le Monde corpus) that contain the two keywords.

This function corresponds to the Contain route of the API.

Examples

if (FALSE) {
  gallicagram_cooccur("président", "mauvais")
}