ngrams that most frequently contain a keyword in a Gallicagram corpus for a specific month

Returns the most frequent ngrams containing a keyword for a given month

Usage

gallicagram_with_month(
  keyword,
  corpus = "lemonde",
  year = 2022,
  month = 1,
  n_results = 20,
  after = FALSE,
  length = 2
)

Arguments

keyword

A character string. Keyword to search. The largest number of words the keyword can contain can be found in the resolution column of the list_corpora dataset.

corpus

A character string. The corpus to search. The list of available corpora can be found in the list_corpora dataset.

year

An integer. The year of interest.

month

An integer. The month of the year of interest.

n_results

An integer. The number of most frequently associated words to return. n_results can also be set to "all" to return all the available results.

after

A boolean. Whether to consider only words following the keyword and not those preceding. Set to FALSE by default.

length

An integer. The length of the ngrams considered. Can be up to 3 in the "books" and "press" corpora and 4 in the "lemonde" corpus.

Value

A tibble. With the n_results most frequent ngrams containing the keyword searched (ngram) and the number of occurrences over the period (n_occur). It also returns the input parameters keyword, corpus, year and month.

Details

This function corresponds to the joker_month route of the API.

It is analogous to gallicagram_with but for a precise month instead of a given period.

For instance "camarade" is often followed by "staline" or "khrouchtchev" in Le Monde. The function returns the most frequent ngrams of the form "camarade *" when setting after = TRUE. after = FALSE also includes the most frequent ngrams of the form "* camarade".

Searching the "press" corpus can require a long running time.

Examples

if (FALSE) {
  gallicagram_with_month("camarade", length = 2)
}